requirements.txt
Django==3.2.16
django-ckeditor==6.7.0
django-js-asset==2.1.0
Pillow==10.0.1
requests~=2.27.1
icalendar~=5.0.11
django-allauth==0.61.1